changing vectore store files structure
