+++ .github/workflows/news.yml
@@ -5,7 +5,7 @@ on:
 
 jobs:
   scrape-news:
-    runs-on: ubuntu-24.04
+    runs-on: ubuntu-22.04
 
     steps:
       - name: Checkout Repository
