+++ src/rag_system/graph_nodes/__init__.py
@@ -4,4 +4,4 @@ from .generator import generation_chain
 from .query_router import question_router
 from .hallucination_grader import hallucination_grader_agent
 from .answer_grader import answer_grader_agent
-from .vectorstore import query_embeddings, doc_embeddings, vector_store
+
