+++ utils/__init__.py
@@ -2,5 +2,5 @@ from .constants import HEADERS, SYMBOLS, RAW_DATA_DOWNLOAD_BASELINK
 from .download_automation import get_dates, update_dates, download_data
 from .train_test_utils import split_dataset, get_target_from_dataset, get_features_from_dataset, train, evaluate, plot_evaluation_result
 from .process_stock_data import process_all_raw_files, fill_missing_dates_interpolation
-from .pinecone_index import get_pinecone_index 
+from .pinecone_vector_store import get_pinecone_vector_store
